pyproject.toml
[project]
name = "oam-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Oppliger J.", email = "account@oppliger.cc" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx-sse>=0.4.3",
    "httpx[http2]>=0.28.1",
    "open-asset-model>=1.1.3",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
